+++ src/main/java/racingcar/engine/Referee.java
@@ -1,4 +1,4 @@
-package racingcar.engine.model;
+package racingcar.engine;
 
 import java.util.List;
 import java.util.stream.Collectors;
