+++ src/test/java/racingcar/InputParserTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.engine.model.InputParser;
+import racingcar.engine.InputParser;
 
 public class InputParserTest {
