+++ src/test/java/racingcar/RefereeTest.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import racingcar.engine.model.Car;
-import racingcar.engine.model.Referee;
+import racingcar.engine.Referee;
 
 public class RefereeTest {
